RR2: Started and Finished
Went ahead and implemented a round-robin scheduling system. We don't
have to use it, I was merely doing it as an exercise.

Added a handful of new variables and functions to gen_proc.cpp.

Added ability to measure context-switch penalty to SJF and FIFO.

I have an idea for simulating multiple cores on FIFO and SJF, so I'll
probably knock those out soon.