SJF_QUAD: Started, but there's a bug...
So, for whatever reason, there's a bug that can cause 1 or 2 of the
"cores" to go idle, and not pick up the next process, even though there
are plenty to choose from. I can't quite figure it out right now, but it
causes times to go back up to non-multicore levels. Fun...