Added First Part of Project 2
I know I'm using your Repo, David, but it's the easiest way to get the
code to Noah. Also, you'll probably be taking this class at some point,
so you can use this as a guide if you so wish.

If you don't wish me to update further, I understand, and I'll make my
own repo for it.