Completed Part 3a and 3b
I completed part 3a and 3b for both malloc() and my_malloc().

Basically, it puts it in a "queue", so to speak, in the sense that it
keeps looping through all the processes until it's done with all of
them. Once it goes through 100 processes that won't run, it declares it
impossible and returns a cycle time of INF. This has yet to actually
happen.

If we were running more than one process at a time, there's a chance
that we could actually have an issue here, but it doesn't state that
anywhere in the problem. I swear to god if he gets us on that I'm going
to cut someone.

Anyway, I moved a lot of stuff around and tried to make main() more
modular.

My_malloc() takes longer than malloc(), and the ones with the queue
generally take slightly longer than those, as expected.